Enable RSpec/DescribeClass
