+++ Library/Homebrew/test/cmd/custom-external-command_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "brew custom-external-command", :integration_test do
+RSpec.describe "brew custom-external-command", :integration_test, type: :system do
   it "is supported" do
     mktmpdir do |path|
       cmd = "custom-external-command-#{rand}"
