+++ Library/Homebrew/test/global_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "brew", :integration_test do
+RSpec.describe Homebrew, :integration_test do
   it "does not invoke `require \"formula\"` at startup" do
     expect { brew "verify-formula-undefined" }
       .to not_to_output.to_stdout
