+++ Library/Homebrew/test/os/mac/mach_spec.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
-RSpec.describe "Mach-O" do
+RSpec.describe MachOShim do
   describe "Pathname tests" do
     specify "fat dylib" do
       pn = dylib_path("fat")
