+++ Library/Homebrew/test/patching_spec.rb
@@ -2,7 +2,7 @@
 
 require "formula"
 
-RSpec.describe "patching" do
+RSpec.describe "patching", type: :system do
   let(:formula_subclass) do
     Class.new(Formula) do
       # These are defined within an anonymous class to avoid polluting the global namespace.
@@ -229,15 +229,3 @@ RSpec.describe "patching" do
     end.to raise_error(BuildError)
   end
 end
-
-__END__
-diff --git a/libexec/NOOP b/libexec/NOOP
-index bfdda4c..e08d8f4 100755
---- a/libexec/NOOP
-+++ b/libexec/NOOP
-@@ -1,2 +1,2 @@
- #!/bin/bash
--echo NOOP
-\ No newline at end of file
-+echo ABCD
-\ No newline at end of file
